flux: reset SIGINT and SIGTERM signal handlers after fork()
Problem: When the spindle flux plugin calls fork(2) to start the
backend, it inherits the existing job shell signal disposition for
SIGTERM and SIGINT. Since the main shell signal handler for these
signals doesn't make sense in the spindle process, this leads to the
backend process ignoring the signals at best, or a possible segfault
at worst. In production this has been observed to leave jobs hanging
in CLEANUP state under Flux when a job is canceled, which results
in SIGTERM terminating the job tasks, the main shell process exits,
but the spindle backend stays active.

Reset SIGTERM and SIGINT signal handlers to their default dispositions
after fork(2) in the flux spindle plugin.