flux: set TMPDIR in job environment if not set
Problem: If TMIDIR is not set in the Flux job environment, Spindle
fails.

The flux plugin already checks for TMPDIR and falls back to /tmp if
missing, but it only sets TMPDIR in the job shell environment in this
case. Also set it in the job environment when TMPDIR is not set.