Fixes runTests failure when specifying locaiton.
The TestVerifier constructor checks for $TMPDIR, $TEMPDIR, and
"/tmp" in turn to determine the cache directory to be used.

In the rest of Spindle, that choice can be overridden by the
"location" parameter.  That parameter has not historically been
communicated through to the testing subsystem, leading to
spurious failures due to directory name mismatches.

Fixed by having the backend code send a "magic" message to the
testing subsystem with the correct location.  This solution works
because backend initialization is guaranteed to complete before the
application could send its first test messages.