Fix exec of script executables when script launched by name
When a script is launched by name, to be found in $PATH, the kernel
implementation of exec substitutes the absolute path to the script as
the argument to the interpreter, but Spindle did not replicate this
behavior, leaving argv[0] as is. This is fixed by replacing argv[0] with
orig_path if argv[0] is not a path. Adds test to verify fix.